refactor_constants
