+++ src/main/java/org/ivan_kropachev/restaurant_voting/model/AbstractBaseEntity.java
@@ -7,11 +7,11 @@ import org.springframework.util.Assert;
 
 import javax.persistence.*;
 
-import static org.ivan_kropachev.restaurant_voting.Constants.START_SEQ;
-
 @MappedSuperclass
 @Access(AccessType.FIELD)
 public abstract class AbstractBaseEntity implements HasId {
+    public static final int START_SEQ = 100000;
+    public static final int NOT_FOUND = 0;
 
     @Id
     @SequenceGenerator(name = "global_seq", sequenceName = "global_seq", allocationSize = 1, initialValue = START_SEQ)
