+++ src/main/java/org/ivan_kropachev/restaurant_voting/model/Menu.java
@@ -10,12 +10,12 @@ import javax.validation.constraints.NotNull;
 import java.time.LocalDate;
 import java.util.List;
 
-import static org.ivan_kropachev.restaurant_voting.Constants.DISH_LIST_STR;
-
 @Cache(usage = CacheConcurrencyStrategy.NONSTRICT_READ_WRITE)
 @Entity
 @Table(name = "menu", uniqueConstraints = {@UniqueConstraint(name = "menu_restaurant_id_date_idx", columnNames = {"restaurant_id", "date"})})
 public class Menu extends AbstractBaseEntity {
+    private static final String DISH_LIST_STR = "[\n{\n\"id\": \"null\",\n\"name\": \"dish1\",\n\"price\": 10\n},\n" +
+            "    {\n\"id\": \"null\",\n\"name\": \"dish2\",\n\"price\": 20\n}\n]";
 
     @NotNull
     @ManyToOne(fetch = FetchType.EAGER)
