+++ src/test/java/org/ivan_kropachev/restaurant_voting/DishTestData.java
@@ -4,11 +4,12 @@ import org.ivan_kropachev.restaurant_voting.model.Dish;
 
 import java.util.List;
 
-import static org.ivan_kropachev.restaurant_voting.Constants.DISH1_ID;
+import static org.ivan_kropachev.restaurant_voting.model.AbstractBaseEntity.START_SEQ;
 
 public class DishTestData {
     public static final MatcherFactory.Matcher<Dish> DISH_MATCHER =
             MatcherFactory.usingIgnoringFieldsComparator(Dish.class, "menu");
+    public static final int DISH1_ID = START_SEQ + 14;
 
     public static final Dish dish1 = new Dish(DISH1_ID, "Hachapuri po imeretinski", 550);
     public static final Dish dish2 = new Dish(DISH1_ID + 1, "Vegan Burger", 150);
