+++ src/test/java/org/ivan_kropachev/restaurant_voting/UserTestData.java
@@ -8,12 +8,13 @@ import java.util.Collections;
 import java.util.Date;
 import java.util.List;
 
-import static org.ivan_kropachev.restaurant_voting.Constants.ADMIN_ID;
-import static org.ivan_kropachev.restaurant_voting.Constants.USER_ID;
+import static org.ivan_kropachev.restaurant_voting.model.AbstractBaseEntity.START_SEQ;
 
 public class UserTestData {
     public static final MatcherFactory.Matcher<User> USER_MATCHER =
             MatcherFactory.usingIgnoringFieldsComparator(User.class, "registered", "password");
+    public static final int ADMIN_ID = START_SEQ + 4;
+    public static final int USER_ID = START_SEQ + 5;
 
     public static final User admin =
             new User(ADMIN_ID, "admin", "admin@gmail.com", "admin", Role.ADMIN);
