+++ src/test/java/org/ivan_kropachev/restaurant_voting/VoteTestData.java
@@ -3,12 +3,18 @@ package org.ivan_kropachev.restaurant_voting;
 import org.ivan_kropachev.restaurant_voting.model.Vote;
 
 import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
 import java.util.List;
 
-import static org.ivan_kropachev.restaurant_voting.Constants.*;
+import static org.ivan_kropachev.restaurant_voting.RestaurantTestData.RESTAURANT1_ID;
+import static org.ivan_kropachev.restaurant_voting.UserTestData.USER_ID;
+import static org.ivan_kropachev.restaurant_voting.model.AbstractBaseEntity.START_SEQ;
 
 public class VoteTestData {
     public static final MatcherFactory.Matcher<Vote> VOTE_MATCHER = MatcherFactory.usingEqualsComparator(Vote.class);
+    public static final int VOTE1_ID = START_SEQ + 23;
+    //https://stackoverflow.com/questions/8746084/string-to-localdate
+    public static final DateTimeFormatter PATTERN = DateTimeFormatter.ofPattern("yyyy-MM-dd");
 
     public static final Vote vote1 =
             new Vote(VOTE1_ID, USER_ID, RESTAURANT1_ID + 1, LocalDate.parse("2021-12-02", PATTERN));
