+++ src/test/java/org/ivan_kropachev/restaurant_voting/service/AbstractRestaurantServiceTest.java
@@ -5,9 +5,8 @@ import org.ivan_kropachev.restaurant_voting.util.exception.NotFoundException;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
-import static org.ivan_kropachev.restaurant_voting.Constants.NOT_FOUND;
-import static org.ivan_kropachev.restaurant_voting.Constants.RESTAURANT1_ID;
 import static org.ivan_kropachev.restaurant_voting.RestaurantTestData.*;
+import static org.ivan_kropachev.restaurant_voting.model.AbstractBaseEntity.NOT_FOUND;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 public class AbstractRestaurantServiceTest extends AbstractServiceTest {
