+++ src/test/java/org/ivan_kropachev/restaurant_voting/web/dish/AdminDishControllerTest.java
@@ -11,11 +11,10 @@ import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 
-import static org.ivan_kropachev.restaurant_voting.Constants.DISH1_ID;
-import static org.ivan_kropachev.restaurant_voting.Constants.NOT_FOUND;
 import static org.ivan_kropachev.restaurant_voting.DishTestData.*;
 import static org.ivan_kropachev.restaurant_voting.TestUtil.userHttpBasic;
 import static org.ivan_kropachev.restaurant_voting.UserTestData.admin;
+import static org.ivan_kropachev.restaurant_voting.model.AbstractBaseEntity.NOT_FOUND;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
